+++ frontend/eui/src/components/drawer/EDrawer.vue
@@ -122,7 +122,7 @@ watch(localOpen, (val) => {
         <DialogClose
           v-if="showClose"
           data-slot="drawer-close"
-          class="ring-offset-background focus:ring-ring absolute top-4 right-4 rounded-xs opacity-70 transition-opacity hover:opacity-100 focus:ring-2 focus:ring-offset-2 focus:outline-hidden disabled:pointer-events-none"
+          class="focus-visible:ring-ring/50 absolute top-4 right-4 rounded-xs opacity-70 transition-opacity hover:opacity-100 focus-visible:ring-[3px] focus-visible:outline-hidden disabled:pointer-events-none"
         >
           <X class="size-4" />
           <span class="sr-only">Close</span>