+++ frontend/eui/src/components/sheet/ESheet.vue
@@ -88,7 +88,7 @@ function onUpdate(value: boolean) {
         <!-- Close button -->
         <DialogClose
           v-if="showClose"
-          class="ring-offset-background focus:ring-ring data-[state=open]:bg-secondary absolute top-4 right-4 rounded-xs opacity-70 transition-opacity hover:opacity-100 focus:ring-2 focus:ring-offset-2 focus:outline-hidden disabled:pointer-events-none"
+          class="focus-visible:ring-ring/50 data-[state=open]:bg-secondary absolute top-4 right-4 rounded-xs opacity-70 transition-opacity hover:opacity-100 focus-visible:ring-[3px] focus-visible:outline-hidden disabled:pointer-events-none"
         >
           <X class="size-4" />
           <span class="sr-only">Close</span>