+++ frontend/eui/src/components/tags-input/ETagsInput.vue
@@ -69,7 +69,7 @@ function onInvalid(payload: string) {
       v-for="item in modelValue"
       :key="item"
       :value="item"
-      class="flex h-5 items-center rounded-md bg-secondary data-[state=active]:ring-ring data-[state=active]:ring-2 data-[state=active]:ring-offset-2 ring-offset-background"
+      class="flex h-5 items-center rounded-md bg-secondary data-[state=active]:ring-ring/50 data-[state=active]:ring-[3px]"
     >
       <TagsInputItemText class="py-0.5 px-2 text-sm rounded bg-transparent" />
       <TagsInputItemDelete class="flex rounded bg-transparent mr-1">